Migrate loan page
