--- resources/js/old/pages/home.vue
@@ -1,27 +1,0 @@
-<template>
-    <div class="row">
-        <div class="col-md-offset-4 col-md-4 text-center" style="margin-top: 10vh">
-            <h1 class="text-light-blue">
-                <i class="fa fa-database fa-5x"></i>
-            </h1>
-            <h1 class="text-light-blue">E M S<br>
-                <small class="text-light-blue">
-                    <strong>E</strong>quipment <strong>M</strong>anagement <strong>S</strong>ystem
-                </small>
-            </h1>
-        </div>
-    </div>
-</template>
-<script>
-    export default {
-        data: function () {
-            return {}
-        },
-        computed: {},
-        methods: {},
-        created: function () {
-
-        },
-        watch: {}
-    }
-</script>
