+++ resources/js/pages/setting/user.vue
@@ -1,5 +1,5 @@
 <template>
-    <div>
+    <div class="container-fluid">
         <h1>
             使用者管理
             <small>Users</small>
